Crate rusoto_secretsmanager

AWS Secrets Manager API Reference
AWS Secrets Manager is a web service that enables you to store, manage, and retrieve, secrets.

This guide provides descriptions of the Secrets Manager API. For more information about using this service, see the AWS Secrets Manager User Guide.

API Version

This version of the Secrets Manager API Reference documents the Secrets Manager API version 2017-10-17.

As an alternative to using the API directly, you can use one of the AWS SDKs, which consist of libraries and sample code for various programming languages and platforms (such as Java, Ruby, .NET, iOS, and Android). The SDKs provide a convenient way to create programmatic access to AWS Secrets Manager. For example, the SDKs take care of cryptographically signing requests, managing errors, and retrying requests automatically. For more information about the AWS SDKs, including how to download and install them, see Tools for Amazon Web Services.

We recommend that you use the AWS SDKs to make programmatic API calls to Secrets Manager. However, you also can use the Secrets Manager HTTP Query API to make direct calls to the Secrets Manager web service. To learn more about the Secrets Manager HTTP Query API, see Making Query Requests in the AWS Secrets Manager User Guide.

Secrets Manager supports GET and POST requests for all actions. That is, the API doesn't require you to use GET for some actions and POST for others. However, GET requests are subject to the limitation size of a URL. Therefore, for operations that require larger sizes, use a POST request.

Support and Feedback for AWS Secrets Manager

We welcome your feedback. Send your comments to awssecretsmanager-feedback@amazon.com, or post your feedback and questions in the AWS Secrets Manager Discussion Forum. For more information about the AWS Discussion Forums, see Forums Help.

How examples are presented

The JSON that AWS Secrets Manager expects as your request parameters and that the service returns as a response to HTTP query requests are single, long strings without line breaks or white space formatting. The JSON shown in the examples is formatted with both line breaks and white space to improve readability. When example input parameters would also result in long strings that extend beyond the screen, we insert line breaks to enhance readability. You should always submit the input as a single JSON text string.

Logging API Requests

AWS Secrets Manager supports AWS CloudTrail, a service that records AWS API calls for your AWS account and delivers log files to an Amazon S3 bucket. By using information that's collected by AWS CloudTrail, you can determine which requests were successfully made to Secrets Manager, who made the request, when it was made, and so on. For more about AWS Secrets Manager and its support for AWS CloudTrail, see Logging AWS Secrets Manager Events with AWS CloudTrail in the AWS Secrets Manager User Guide. To learn more about CloudTrail, including how to turn it on and find your log files, see the AWS CloudTrail User Guide.

If you're using the service, you're probably looking for SecretsManagerClient and SecretsManager.

Structs

CancelRotateSecretRequest

CancelRotateSecretResponse

CreateSecretRequest

CreateSecretResponse

DeleteResourcePolicyRequest

DeleteResourcePolicyResponse

DeleteSecretRequest

DeleteSecretResponse

DescribeSecretRequest

DescribeSecretResponse

GetRandomPasswordRequest

GetRandomPasswordResponse

GetResourcePolicyRequest

GetResourcePolicyResponse

GetSecretValueRequest

GetSecretValueResponse

ListSecretVersionIdsRequest

ListSecretVersionIdsResponse

ListSecretsRequest

ListSecretsResponse

PutResourcePolicyRequest

PutResourcePolicyResponse

PutSecretValueRequest

PutSecretValueResponse

RestoreSecretRequest

RestoreSecretResponse

RotateSecretRequest

RotateSecretResponse

RotationRulesType
A structure that defines the rotation configuration for the secret.

SecretListEntry
A structure that contains the details about a secret. It does not include the encrypted SecretString and SecretBinary values. To get those values, use the GetSecretValue operation.

SecretVersionsListEntry
A structure that contains information about one version of a secret.

SecretsManagerClient
A client for the AWS Secrets Manager API.

Tag
A structure that contains information about a tag.

TagResourceRequest

UntagResourceRequest

UpdateSecretRequest

UpdateSecretResponse

UpdateSecretVersionStageRequest

UpdateSecretVersionStageResponse

Enums

CancelRotateSecretError
Errors returned by CancelRotateSecret

CreateSecretError
Errors returned by CreateSecret

DeleteResourcePolicyError
Errors returned by DeleteResourcePolicy

DeleteSecretError
Errors returned by DeleteSecret

DescribeSecretError
Errors returned by DescribeSecret

GetRandomPasswordError
Errors returned by GetRandomPassword

GetResourcePolicyError
Errors returned by GetResourcePolicy

GetSecretValueError
Errors returned by GetSecretValue

ListSecretVersionIdsError
Errors returned by ListSecretVersionIds

ListSecretsError
Errors returned by ListSecrets

PutResourcePolicyError
Errors returned by PutResourcePolicy

PutSecretValueError
Errors returned by PutSecretValue

RestoreSecretError
Errors returned by RestoreSecret

RotateSecretError
Errors returned by RotateSecret

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateSecretError
Errors returned by UpdateSecret

UpdateSecretVersionStageError
Errors returned by UpdateSecretVersionStage

Traits

SecretsManager
Trait representing the capabilities of the AWS Secrets Manager API. AWS Secrets Manager clients implement this trait.