Struct rusoto_secretsmanager::SecretsManagerClient

pub struct SecretsManagerClient { /* fields omitted */ }

A client for the AWS Secrets Manager API.

Methods

impl SecretsManagerClient

pub fn new(region: Region) -> SecretsManagerClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> SecretsManagerClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,