Struct rusoto_secretsmanager::UpdateSecretRequest

pub struct UpdateSecretRequest {
    pub client_request_token: Option<String>,
    pub description: Option<String>,
    pub kms_key_id: Option<String>,
    pub secret_binary: Option<Vec<u8>>,
    pub secret_id: String,
    pub secret_string: Option<String>,
}

Fields
client_request_token: Option<String>

(Optional) If you want to add a new version to the secret, this parameter specifies a unique identifier for the new version that helps ensure idempotency.

If you use the AWS CLI or one of the AWS SDK to call this operation, then you can leave this parameter empty. The CLI or SDK generates a random UUID for you and includes that in the request. If you don't use the SDK and instead generate a raw HTTP request to the AWS Secrets Manager service endpoint, then you must generate a ClientRequestToken yourself for new versions and include that value in the request.

You typically only need to interact with this value if you implement your own retry logic and want to ensure that a given secret is not created twice. We recommend that you generate a UUID-type value to ensure uniqueness within the specified secret.

Secrets Manager uses this value to prevent the accidental creation of duplicate versions if there are failures and retries during the Lambda rotation function's processing.

If the ClientRequestToken value isn't already associated with a version of the secret then a new version of the secret is created.

If a version with this value already exists and that version's SecretString and SecretBinary values are the same as those in the request then the request is ignored (the operation is idempotent).

If a version with this value already exists and that version's SecretString and SecretBinary values are different from the request then an error occurs because you cannot modify an existing secret value.

This value becomes the SecretVersionId of the new version.

description: Option<String>

(Optional) Specifies a user-provided description of the secret.

kms_key_id: Option<String>

(Optional) Specifies the ARN or alias of the KMS customer master key (CMK) to be used to encrypt the protected text in the versions of this secret.

If you don't specify this value, then Secrets Manager defaults to using the default CMK in the account (the one named aws/secretsmanager). If a KMS CMK with that name doesn't exist, then AWS Secrets Manager creates it for you automatically the first time it needs to encrypt a version's Plaintext or PlaintextString fields.

You can only use the account's default CMK to encrypt and decrypt if you call this operation using credentials from the same account that owns the secret. If the secret is in a different account, then you must create a custom CMK and provide the ARN in this field.

secret_binary: Option<Vec<u8>>

(Optional) Specifies binary data that you want to encrypt and store in the new version of the secret. To use this parameter in the command-line tools, we recommend that you store your binary data in a file and then use the appropriate technique for your tool to pass the contents of the file as a parameter. Either SecretBinary or SecretString must have a value. They cannot both be empty.

This parameter is not accessible using the Secrets Manager console.

secret_id: String

Specifies the secret that you want to update or to which you want to add a new version. You can specify either the Amazon Resource Name (ARN) or the friendly name of the secret.

secret_string: Option<String>

(Optional) Specifies text data that you want to encrypt and store in this new version of the secret. Either SecretBinary or SecretString must have a value. They cannot both be empty.

If you create this secret by using the Secrets Manager console then Secrets Manager puts the protected secret text in only the SecretString parameter. The Secrets Manager console stores the information as a JSON structure of key/value pairs that the default Lambda rotation function knows how to parse.

For storing multiple values, we recommend that you use a JSON text string argument and specify key/value pairs. For information on how to format a JSON parameter for the various command line tool environments, see Using JSON for Parameters in the AWS CLI User Guide.